make common test hooks module
