+++ tests/tooling/hooks.py
@@ -1,0 +1,11 @@
+from idom import use_state
+
+
+def use_toggle(init=False):
+    state, set_state = use_state(init)
+    return state, lambda: set_state(lambda old: not old)
+
+
+def use_counter(initial_value):
+    state, set_state = use_state(initial_value)
+    return state, lambda: set_state(state + 1)
